requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.5
django-allauth==0.44.0
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-registration==3.2
django-rest-auth==0.9.5
django-templated-mail==1.1.1
django-webpack-loader==1.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
drf-writable-nested==0.6.3
flake8==3.9.2
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.1
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
